+++ setup_gpl.py
@@ -25,6 +25,9 @@ setuptools.setup(
     license="GNU General Public License v3",
     packages=["miss_hit_core"],
     python_requires=">=3.6, <4",
+    package_data = {
+        "miss_hit_core": ["resources/*.css", "resources/assets/*"],
+    },
     classifiers=[
         "Development Status :: 4 - Beta",
         "Intended Audience :: Developers",